deepzero/requirements.txt
numpy>=1.23
cupy>=11.0
pyyaml>=6.0
tqdm>=4.64
tensorboard>=2.12